move file/watcher state into InstanceState, flatten service facades, delete runtime
- File service: InstanceState for cache/fiber, flatten index.ts + service.ts
- FileWatcher: wrap subscription setup in InstanceState, add init()
- Question: flatten index.ts + service.ts
- ProviderAuth: flatten auth.ts + auth-service.ts, remove fn() wrappers
- Account CLI: use Account.runPromise instead of runtime
- Delete effect/runtime.ts and effect/instances.ts (no more centralized runtime)
- Delete runtime.test.ts (no longer applicable)
