+++ packages/opencode/src/effect/run-service.ts
@@ -6,7 +6,7 @@ export const memoMap = Layer.makeMemoMapUnsafe()
 export function makeRunPromise<I, S, E>(service: ServiceMap.Service<I, S>, layer: Layer.Layer<I, E>) {
   let rt: ManagedRuntime.ManagedRuntime<I, E> | undefined
 
-  return <A, Err>(fn: (svc: S) => Effect.Effect<A, Err>, options?: Effect.RunOptions) => {
+  return <A, Err>(fn: (svc: S) => Effect.Effect<A, Err, I>, options?: Effect.RunOptions) => {
     rt ??= ManagedRuntime.make(layer, { memoMap })
     return rt.runPromise(service.use(fn), options)
   }
