+++ packages/opencode/src/project/vcs.ts
@@ -97,7 +97,7 @@ export namespace Vcs {
         }),
       })
     }),
-  ).pipe(Layer.fresh)
+  )
 }
 
 export const runPromise = makeRunPromise(Vcs.Service, Vcs.layer)
